# contracts/openapi/local-api.json
{
  "components": {
    "schemas": {
...
      "CreateEvaluationSetRequest": {
        "additionalProperties": false,
        "properties": {
          "anchors": {
            "items": {
              "$ref": "#/components/schemas/EvaluationAnchorInput"
            },
            "title": "Anchors",
            "type": "array"
          },
          "configurations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationConfigurationInput"
            },
            "title": "Configurations",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "evaluator_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evaluator Ids",
            "type": "array"
          },
          "hidden_metadata_policy": {
            "additionalProperties": true,
            "title": "Hidden Metadata Policy",
            "type": "object"
          },
          "library_id": {
            "minLength": 1,
            "title": "Library Id",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "maxLength": 1000,
            "title": "Purpose",
            "type": "string"
          },
          "split_policy": {
            "additionalProperties": true,
            "title": "Split Policy",
            "type": "object"
          },
          "track_ids": {
            "items": {
              "type": "string"
            },
            "title": "Track Ids",
            "type": "array"
          },
          "track_splits": {
            "items": {
              "$ref": "#/components/schemas/TrackSplit"
            },
            "title": "Track Splits",
            "type": "array"
          },
          "version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "library_id",
          "name",
          "version"
        ],
        "title": "CreateEvaluationSetRequest",
        "type": "object"
      },
...
      "EvaluationNeighborsResponse": {
        "properties": {
          "anchor_track_id": {
            "title": "Anchor Track Id",
            "type": "string"
          },
          "configurations": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Configurations",
            "type": "array"
          },
          "contract_version": {
            "const": "evaluation.v1",
            "title": "Contract Version",
            "type": "string"
          },
          "evaluation_run_id": {
            "title": "Evaluation Run Id",
            "type": "string"
          },
          "evaluation_set_id": {
            "title": "Evaluation Set Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "evaluation_set_id",
          "evaluation_run_id",
          "anchor_track_id",
          "configurations"
        ],
        "title": "EvaluationNeighborsResponse",
        "type": "object"
      },
...
      "EvaluationSetDetailResponse": {
        "properties": {
          "anchors": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Anchors",
            "type": "array"
          },
          "configurations": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Configurations",
            "type": "array"
          },
          "contract_version": {
            "const": "evaluation.v1",
            "title": "Contract Version",
            "type": "string"
          },
          "evaluation_set": {
            "$ref": "#/components/schemas/EvaluationSetResponse"
          },
          "latest_run": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run"
          },
          "tracks": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "evaluation_set",
          "tracks",
          "anchors",
          "configurations",
          "latest_run"
        ],
        "title": "EvaluationSetDetailResponse",
        "type": "object"
      },
      "EvaluationSetListResponse": {
        "properties": {
          "contract_version": {
            "const": "evaluation.v1",
            "title": "Contract Version",
            "type": "string"
          },
          "evaluation_sets": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSetResponse"
            },
            "title": "Evaluation Sets",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "evaluation_sets"
        ],
        "title": "EvaluationSetListResponse",
        "type": "object"
      },
      "EvaluationSetResponse": {
        "additionalProperties": true,
        "properties": {
          "anchor_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Count"
          },
          "configuration_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Count"
          },
          "configurations": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Configurations",
            "type": "array"
          },
          "contract_version": {
            "const": "evaluation.v1",
            "title": "Contract Version",
            "type": "string"
          },
          "hidden_metadata": {
            "title": "Hidden Metadata",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "track_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Count"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "id",
          "library_id",
          "name",
          "version",
          "hidden_metadata"
        ],
        "title": "EvaluationSetResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
    }
  },
  "info": {
    "title": "Crate Dig local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluation-sets": {
      "get": {
        "operationId": "list_sets_evaluation_sets_get",
        "parameters": [
          {
            "in": "query",
            "name": "library_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sets",
        "tags": [
          "similarity-evaluation"
        ]
      },
      "post": {
        "operationId": "create_set_evaluation_sets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Set",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}": {
      "get": {
        "operationId": "get_set_evaluation_sets__evaluation_set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Set",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}/anchors/{anchor_track_id}/neighbors": {
      "get": {
        "operationId": "get_neighbors_evaluation_sets__evaluation_set_id__anchors__anchor_track_id__neighbors_get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "anchor_track_id",
            "required": true,
            "schema": {
              "title": "Anchor Track Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "configuration_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Configuration Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationNeighborsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neighbors",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}/judgments": {
      "post": {
        "operationId": "create_judgment_evaluation_sets__evaluation_set_id__judgments_post",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJudgmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Judgment",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}/next": {
      "get": {
        "operationId": "get_next_evaluation_sets__evaluation_set_id__next_get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "evaluator_id",
            "required": false,
            "schema": {
              "default": "local",
              "title": "Evaluator Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "anchor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anchor Id"
            }
          },
          {
            "in": "query",
            "name": "configuration_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Configuration Id"
            }
          },
          {
            "in": "query",
            "name": "configuration_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Configuration Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRoundResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Next",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}/report": {
      "get": {
        "operationId": "report_evaluation_sets__evaluation_set_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
    "/evaluation-sets/{evaluation_set_id}/runs": {
      "post": {
        "operationId": "create_run_evaluation_sets__evaluation_set_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_set_id",
            "required": true,
            "schema": {
              "title": "Evaluation Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "similarity-evaluation"
        ]
      }
    },
...
  }
}